To build a Toffoli gate with base set.
The problem
We have seen above how to construct a Hadamard gate with our base set, and now we want to build a Toffoli gate as well. Why the Toffoli gate? As mentioned above, the Toffoli gate is also a universal gate for classical computation the same way the NAND gate is, but it is reversible. Further it builds a simple universal gate set for quantum computation if combined with the Hadamard gate.

We have seen some examples on how to express more complex gates using basis gates, we now want to use the knowledge gained to construct a Toffoli gate only using our basis gates. In order to do solve this exercise the above examples on how to construct and use controlled rotations, will come in handy. The biggest challenge is to construct the needed controlled rotations.

You can use the code below using the composer widget to construct your circuit.

As a reminder the basis gates for IBM Quantum systems are CX, RZ, SX and X gates, so no other gates are allowed.

Of course we want also try to minimize the cost.

𝐶𝑜𝑠𝑡=10𝑁𝐶𝑁𝑂𝑇+𝑁𝑜𝑡ℎ𝑒𝑟